refactor to params dict
